guessing-micro-full/services/score-service/requirements.txt
fastapi
uvicorn[standard]
flask
psycopg2-binary

